chore: Make GIZMO API methods static with overload for Quarkus
