+++ core/core-impl/src/test/java/ai/timefold/solver/core/impl/domain/solution/cloner/gizmo/GizmoSolutionClonerTest.java
@@ -69,7 +69,7 @@ class GizmoSolutionClonerTest extends AbstractSolutionClonerTest {
                             generateGizmoSolutionOrEntityDescriptor(solutionDescriptor, clazz));
                 });
 
-        new GizmoSolutionClonerImplementor().defineClonerFor(classCreator, solutionDescriptor,
+        GizmoSolutionClonerImplementor.defineClonerFor(classCreator, solutionDescriptor,
                 Collections.singleton(solutionDescriptor.getSolutionClass()),
                 memoizedSolutionOrEntityDescriptorMap, deepClonedClassSet);
         classCreator.close();
